♻️refactor : 재사용성이 있는 함수 utils로 이동
